third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-domain.html

<!DOCTYPE html>
<html>
<head>
<script>
</script>
<link id="target" rel="import" href="resources/having-domain-policy.html">
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBe("document.inlineScriptHasRun", "undefined");
shouldBe("document.externalScriptHasRun", "undefined");
shouldBeTrue("document.corsExternalScriptHasRun");
shouldBe("document.evalFromInlineHasRun", "undefined");
shouldBe("document.evalFromExternalHasRun", "undefined");
shouldBe("document.evalFromCorsExternalHasRun", "undefined");
</script>
</body>
</html>
